Move e3.aws.Stack definition to e3.aws.troposphere.Stack
This avoid dependency on e3.aws.cfn in e3.aws module
