+++ src/e3/aws/troposphere/config/config_rule.py
@@ -8,7 +8,8 @@ if TYPE_CHECKING:
 
 from troposphere import AWSObject, config
 
-from e3.aws import Construct, name_to_id
+from e3.aws import name_to_id
+from e3.aws.troposphere import Construct
 
 
 @dataclass(frozen=True)
