+++ src/e3/aws/troposphere/config/configuration_recorder.py
@@ -9,7 +9,8 @@ if TYPE_CHECKING:
 
 from troposphere import AWSObject, AccountId, iam, config, Join
 
-from e3.aws import Construct, name_to_id
+from e3.aws import name_to_id
+from e3.aws.troposphere import Construct
 from e3.aws.troposphere.s3.bucket import AWSConfigBucket
