+++ src/e3/aws/troposphere/ecr/repository.py
@@ -9,7 +9,8 @@ if TYPE_CHECKING:
 
 from troposphere import AWSObject, ecr, Tags
 
-from e3.aws import Construct, name_to_id
+from e3.aws import name_to_id
+from e3.aws.troposphere import Construct
 
 
 @dataclass(frozen=True)
