+++ src/e3/aws/troposphere/ecs/cluster.py
@@ -7,7 +7,8 @@ if TYPE_CHECKING:
 
 from troposphere import AWSObject, ecs, GetAtt, Ref, Tags
 
-from e3.aws import Construct, name_to_id
+from e3.aws import name_to_id
+from e3.aws.troposphere import Construct
 from e3.aws.troposphere.iam.role import Role
 from e3.aws.troposphere.iam.managed_policy import ManagedPolicy
 from e3.aws.troposphere.iam.policy_statement import PolicyStatement
