+++ src/e3/aws/troposphere/iam/role.py
@@ -9,7 +9,8 @@ if TYPE_CHECKING:
 
 from troposphere import AWSObject, iam, Tags
 
-from e3.aws import Construct, name_to_id
+from e3.aws import name_to_id
+from e3.aws.troposphere import Construct
 from e3.aws.troposphere.iam.policy_document import PolicyDocument
 from e3.aws.troposphere.iam.policy_statement import AssumeRole
